Request-ID plugin adds Snowflake algorithm
